+++ experiments/crydrview/CrydrViewERC20Validatable.sol
@@ -17,7 +17,7 @@ contract CrydrViewERC20Validatable is CrydrViewBase,
 
   /* Constructor */
 
-  function CrydrViewERC20Validatable(string _name, string _symbol, uint32 _decimals)
+  function CrydrViewERC20Validatable(string _name, string _symbol, uint8 _decimals)
     CrydrViewBase('erc20__validatable', 0xffffffff)
     ERC20Named(_name, _symbol, _decimals) {}
